Correct scatterplot directory and add id
