+++ src/app/+data-and-analytics/frequency-graph/scatterplot/scatterplot.component.spec.ts
