--- src/app/+data-and-analytics/frequency-graph/scattorplot/scatterplot.component.ts
@@ -1,11 +1,0 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
-
-@Component({
-  selector: 'scholars-scatterplot',
-  templateUrl: './scatterplot.component.html',
-  styleUrls: ['./scatterplot.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush,
-})
-export class ScatterplotComponent {
-
-}
